+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
-## Unreleased
+## [Unreleased]
+
+## [v0.2.0] - 2021-03-13
 
 ### Added
 
@@ -28,6 +30,7 @@
 - Untracking of files.
 - Tracking registry initialization.
 
-[unreleased]: https://github.com/slok/agebox/compare/v0.1.1...HEAD
+[unreleased]: https://github.com/slok/agebox/compare/v0.2.0...HEAD
+[v0.2.0]: https://github.com/slok/agebox/compare/v0.1.1...v0.2.0
 [v0.1.1]: https://github.com/slok/agebox/compare/v0.1.0...v0.1.1
 [v0.1.0]: https://github.com/slok/agebox/releases/tag/v0.1.0